requirements.txt
dj-database-url==0.4.0
django==1.7.1
django_selectable==0.9.0
gunicorn==19.4.5
whitenoise==2.0.6
psycopg2==2.6.1
